Extract Query class from Benchmark into its own top-level class.
